prepare for theme extending
