+++ dev/package.json
@@ -9,12 +9,10 @@
   },
   "dependencies": {
     "@kobalte/solidbase": "workspace:*",
-    "@solidjs/router": "^0.14.7",
-    "@solidjs/start": "^1.0.8",
-    "solid-js": "^1.9.1",
-    "vinxi": "^0.4.3",
-    "@vinxi/plugin-mdx": "^3.7.2",
-    "solid-mdx": "^0.0.7"
+    "@solidjs/router": "^0.14.10",
+    "@solidjs/start": "^1.0.9",
+    "solid-js": "^1.9.3",
+    "vinxi": "^0.4.3"
   },
   "engines": {
     "node": ">=18"
