+++ docs/package.json
@@ -9,10 +9,10 @@
   },
   "dependencies": {
     "@kobalte/solidbase": "workspace:*",
-    "@solidjs/router": "^0.14.7",
-    "@solidjs/start": "^1.0.8",
+    "@solidjs/router": "^0.14.10",
+    "@solidjs/start": "^1.0.9",
     "@vinxi/plugin-mdx": "^3.7.2",
-    "solid-js": "^1.9.1",
+    "solid-js": "^1.9.3",
     "solid-mdx": "^0.0.7",
     "vinxi": "^0.4.3"
   },
