+++ package.json
@@ -33,6 +33,11 @@
       "import": "./src/server.ts",
       "types": "./src/server.ts"
     },
+    "./default-theme": {
+      "solid": "./src/default-theme/index.ts",
+      "import": "./src/default-theme/index.ts",
+      "types": "./src/default-theme/index.ts"
+    },
     "./*": {
       "solid": "./src/*",
       "import": "./src/*",
@@ -40,14 +45,14 @@
     }
   },
   "devDependencies": {
-    "@biomejs/biome": "^1.9.2",
-    "@solidjs/start": "^1.0.8",
+    "@biomejs/biome": "^1.9.4",
+    "@solidjs/start": "^1.0.9",
     "@types/cross-spawn": "^6.0.6",
     "prettier": "4.0.0-alpha.8",
-    "solid-js": "^1.9.1",
-    "typescript": "^5.6.2",
+    "solid-js": "^1.9.3",
+    "typescript": "^5.6.3",
     "vinxi": "^0.4.3",
-    "vite": "^5.4.8"
+    "vite": "^5.4.10"
   },
   "publishConfig": {
     "access": "public"
@@ -62,10 +67,10 @@
   "dependencies": {
     "@docsearch/css": "^3.6.2",
     "@docsearch/js": "^3.6.2",
-    "@expressive-code/plugin-collapsible-sections": "^0.37.0",
-    "@expressive-code/plugin-line-numbers": "^0.37.0",
+    "@expressive-code/plugin-collapsible-sections": "^0.37.1",
+    "@expressive-code/plugin-line-numbers": "^0.37.1",
     "@kobalte/core": "^0.13.7",
-    "@mdx-js/mdx": "^3.0.1",
+    "@mdx-js/mdx": "^3.1.0",
     "@solid-primitives/clipboard": "^1.5.10",
     "@solid-primitives/context": "^0.2.3",
     "@solid-primitives/event-listener": "^2.3.3",
@@ -74,7 +79,7 @@
     "@solid-primitives/platform": "^0.1.2",
     "@solid-primitives/scroll": "^2.0.23",
     "@solidjs/meta": "^0.29.4",
-    "@solidjs/router": "^0.14.7",
+    "@solidjs/router": "^0.14.10",
     "@vinxi/plugin-mdx": "^3.7.2",
     "cross-spawn": "^7.0.3",
     "esast-util-from-js": "^2.0.1",
@@ -84,7 +89,7 @@
     "mdast-util-find-and-replace": "^3.0.1",
     "mdast-util-toc": "^7.1.0",
     "rehype-autolink-headings": "^7.1.0",
-    "rehype-expressive-code": "^0.37.0",
+    "rehype-expressive-code": "^0.37.1",
     "rehype-raw": "^7.0.0",
     "rehype-slug": "^6.0.0",
     "remark-directive": "^3.0.0",
