+++ src/config/vite-plugin/index.ts
