+++ src/default-theme/index.css
@@ -1,44 +1,6 @@
 @import "./reset.css";
 @import "./variables.css";
 
-@font-face {
-  font-family: Lexend;
-  font-style: normal;
-  font-weight: 100 900;
-  font-display: swap;
-  src: url("./fonts/Lexend.woff2") format("woff2");
-}
-
-@font-face {
-  font-family: Inter;
-  font-style: normal;
-  font-weight: 100 900;
-  font-display: swap;
-  src: url("./fonts/Inter.woff2") format("woff2");
-}
-@font-face {
-  font-family: InterVariable;
-  font-style: italic;
-  font-weight: 100 900;
-  font-display: swap;
-  src: url("./fonts/Inter-Italic.woff2") format("woff2");
-}
-
-@font-face {
-  font-family: "JetBrains Mono";
-  font-style: normal;
-  font-weight: 100 900;
-  font-display: swap;
-  src: url("./fonts/JetBrainsMono.woff2") format("woff2");
-}
-@font-face {
-  font-family: "JetBrains Mono";
-  font-style: italic;
-  font-weight: 100 900;
-  font-display: swap;
-  src: url("./fonts/JetBrainsMono-Italic.woff2") format("woff2");
-}
-
 html {
   --font-headings: Lexend, Segoe UI, Roboto, sans-serif;
   --font-text: Inter, Segoe UI, Roboto, sans-serif;
