+++ src/default-theme/utils.ts
@@ -1,6 +1,6 @@
 import { useSolidBaseContext as _useSolidBaseContext } from "../client/context";
 import { useRouteConfig as _useRouteConfig } from "../client/config";
-import { DefaultThemeConfig } from "./config";
+import { DefaultThemeConfig } from ".";
 
 export function useSolidBaseContext() {
   return _useSolidBaseContext<DefaultThemeConfig>();
